Initial support for flagr backend and full redesign.
The original design grouped flags by project and environment. This idea is convenient but mildly flawed in that project and environment limitations are merely constraints on how a flag should be returned.

flagr is another feature flag service that provides a robust backend but lacks in frontend features like RBAC and real-time flag updates via websockets.

As of this commit, cascade outsources flag storage and evaluation to flagr and provides the missing features as a frontend.

So far, this commit is a POC. You can build a flag in flagr and subscribe to it in cascade from a browser. Subsequent changes must be made in flagr and then a notification must be triggered in cascade.